[babel-plugin] refactor generateCSSRules logic
